refactor: move doInitialSchematicTrace out of Trace.ts
